chore: add new hook for txn logic
added all logic for txn inside the hook for
- erc20-deploy.tsx
- erc20-write-mint.tsx
- erc20-write-transfer.tsx

Tested on =>
- Sepolia Testnet
