+++ integrations/erc20/components/erc20-write-transfer.tsx
@@ -1,7 +1,4 @@
-import { zodResolver } from '@hookform/resolvers/zod'
-import { utils } from 'ethers'
-import { useForm } from 'react-hook-form'
-import { useSigner } from 'wagmi'
+import { ethers } from 'ethers'
 import { z } from 'zod'
 
 import { WalletConnect } from '@/components/blockchain/wallet-connect'
@@ -12,38 +9,21 @@ import { FormDescription, FormItem } from '@/components/ui/form'
 import { Input } from '@/components/ui/input'
 
 import ERC20EventTransfer from './erc20-event-transfer'
-import { useErc20Transfer } from '../erc20-wagmi'
-import { useTokenStorage } from '../hooks/use-token-storage'
+import { useWriteTransfer } from '../hooks/use-write-transfer'
 import { writeTransferControls } from '../utils/controls'
-import { writeTransferFormSchema } from '../utils/formSchema'
+
+const writeTransferFormSchema = z.object({
+  amount: z.string().min(1),
+  fromAddress: z.string().refine((value) => ethers.utils.isAddress(value), {
+    message: 'Sender address is invalid. Please insure you have typed correctly.',
+  }),
+  toAddress: z.string().refine((value) => ethers.utils.isAddress(value), {
+    message: 'Reciever address is invalid. Please insure you have typed correctly.',
+  }),
+})
 
 export function ERC20ContractTransferTokens() {
-  const form = useForm<z.infer<typeof writeTransferFormSchema>>({
-    resolver: zodResolver(writeTransferFormSchema),
-    defaultValues: {
-      fromAddress: '',
-      toAddress: '',
-      amount: '',
-    },
-  })
-  const { data: signer } = useSigner()
-
-  const [token] = useTokenStorage()
-  // @ts-ignore
-  const mintAction = useErc20Transfer({
-    address: token,
-  })
-
-  const onSubmit = async (values: z.infer<typeof writeTransferFormSchema>) => {
-    console.log('ij', values)
-
-    // @ts-ignore
-    const tx = await mintAction.writeAsync({
-      recklesslySetUnpreparedArgs: [values?.toAddress as `0x${string}`, utils.parseEther(values.amount)],
-    })
-
-    form.reset()
-  }
+  const { form, onSubmit } = useWriteTransfer({ writeTransferFormSchema })
 
   return (
     <>
