+++ integrations/erc20/utils/controls.ts
@@ -11,6 +11,16 @@ export const writeMintControls: writeMintType[] = [
 
 export const writeTransferControls: writeTransferType[] = [
   { formfieldName: 'amount', label: 'Amount', placeholder: 'Amount', description: 'This is the amount of the mintable.' },
-  { formfieldName: 'fromAddress', label: 'fromAddress', placeholder: 'fromAddress', description: 'This is the fromAddress of the mintable.' },
-  { formfieldName: 'toAddress', label: 'toAddress', placeholder: 'toAddress', description: 'This is the toAddress of the mintable.' },
+  {
+    formfieldName: 'fromAddress',
+    label: 'Sender Address',
+    placeholder: 'Type Sender Address',
+    description: 'This is the Sender Address of the mintable.',
+  },
+  {
+    formfieldName: 'toAddress',
+    label: 'Receiver Address',
+    placeholder: 'Type Receiver Address',
+    description: 'This is the Receiver Address of the mintable.',
+  },
 ]
